Add currency support to d3.format. Fixes #777.
You can now prefix the locale’s currency symbol. For example:

  d3.format("+$,.2f")(-2.5e5) // -$250,000.00

This implementation is limited in that it does not support currencies where the
symbol should appear after the value (e.g., "250,000.00 €" as in fr_FR). And
d3.format currently only supports a single locale. But it’s a start.